sql/demo/requirements.txt
google-cloud-storage==2.16.0
pandas==2.2.1
pymysql==1.1.0
SQLAlchemy==2.0.29
cloud-sql-python-connector==1.9.1